adding product entity
