+++ src/main/java/com/example/git/demo/pack1/product.java
@@ -1,0 +1,31 @@
+package com.example.git.demo.pack1;
+
+public class product {
+    private int id;
+    private int price;
+    private String name;
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getPrice() {
+        return price;
+    }
+
+    public void setPrice(int price) {
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+}
