--- src/main/java/com/example/git/demo/product.java
@@ -1,42 +1,0 @@
-package com.example.git.demo;
-
-public class product {
-    int id;
-    String name;
-    int price;
-
-    public int getQuantiy() {
-        return quantiy;
-    }
-
-    public void setQuantiy(int quantiy) {
-        this.quantiy = quantiy;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public int getPrice() {
-        return price;
-    }
-
-    public void setPrice(int price) {
-        this.price = price;
-    }
-
-    int quantiy;
-
-}
